+++ app/models/releaf/node.rb
@@ -20,7 +20,6 @@ module Releaf
     accepts_nested_attributes_for :content
 
     after_save :update_settings_timestamp
-    after_validation :run_custom_validations
 
     acts_as_url :name, url_attribute: :slug, scope: :parent_id, :only_when_blank => true
 
@@ -151,16 +150,7 @@ module Releaf
       self_and_ancestors.where(active: false).exists? == false
     end
 
-    def custom_validators
-      content_class.try(:acts_as_node_configuration).try(:[], :validators)
-    end
-
     private
-
-    def run_custom_validations
-      return if custom_validators.blank?
-      self.validates_with *custom_validators
-    end
 
     def update_settings_timestamp
       Settings['nodes.updated_at'] = Time.now